# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatBot Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VerifyCodeInput {
  phone: string;
  code: string;
}

export interface VerifyCodeResult {
  valid: boolean;
  phone: string;
  platformUnlocked: boolean;
  unlockedBots: string[];
  sessionToken?: string;
}

export interface VerifyBotCodeInput {
  phone: string;
  botId: string;
  code: string;
}

export interface VerifyBotCodeResult {
  valid: boolean;
  phone: string;
  botId: string;
  platformUnlocked: boolean;
  unlockedBots: string[];
  sessionToken?: string;
}

export interface UserSession {
  phone: string;
  platformUnlocked: boolean;
  unlockedBots: string[];
  hasAccessCode: boolean;
}

export interface Bot {
  _id: string;
  name: string;
  age: number;
  tagline: string;
  bio?: string;
  avatar: string;
  interests?: string[];
  isOnline: boolean;
}

export interface BotInput {
  name: string;
  age: number;
  tagline: string;
  bio?: string;
  avatar: string;
  interests?: string[];
  scripts?: string[];
}

export interface PaymentInitiateInput {
  phone: string;
}

export interface BotPaymentInput {
  phone: string;
  botId: string;
}

export type MpesaCallbackPayloadBody = { [key: string]: unknown };

export interface MpesaCallbackPayload {
  Body?: MpesaCallbackPayloadBody;
}

export interface PaymentInitiateResult {
  message: string;
  checkoutRequestId: string;
  merchantRequestId?: string;
  transactionRef?: string;
}

export type PaymentStatusPaymentStatus =
  (typeof PaymentStatusPaymentStatus)[keyof typeof PaymentStatusPaymentStatus];

export const PaymentStatusPaymentStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
  cancelled: "cancelled",
  timeout: "timeout",
  none: "none",
} as const;

export interface PaymentStatus {
  platformUnlocked: boolean;
  unlockedBots: string[];
  paymentStatus: PaymentStatusPaymentStatus;
  resultCode?: string;
  resultDesc?: string;
  mpesaReceiptNumber?: string;
  transactionRef?: string;
  merchantRequestId?: string;
  amount?: number;
  completedAt?: string;
  sessionToken?: string;
}

export type TransactionRecordType =
  (typeof TransactionRecordType)[keyof typeof TransactionRecordType];

export const TransactionRecordType = {
  platform: "platform",
  bot: "bot",
} as const;

export type TransactionRecordStatus =
  (typeof TransactionRecordStatus)[keyof typeof TransactionRecordStatus];

export const TransactionRecordStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
  cancelled: "cancelled",
  timeout: "timeout",
} as const;

export interface TransactionRecord {
  _id: string;
  type: TransactionRecordType;
  botId?: string | null;
  amount: number;
  amountCents: number;
  status: TransactionRecordStatus;
  transactionRef?: string | null;
  checkoutRequestId: string;
  merchantRequestId?: string | null;
  mpesaReceiptNumber?: string | null;
  mpesaTransactionDate?: string | null;
  resultCode?: string | null;
  resultDesc?: string | null;
  completedAt?: string | null;
  createdAt: string;
}

export interface PaymentHistoryResult {
  transactions: TransactionRecord[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AdminUnlockResult {
  phone: string;
  accessCode: string;
  message: string;
}

export interface AdminBotUnlockResult {
  phone: string;
  botId: string;
  accessCode: string;
  message: string;
}

export interface BotAccessCodeEntry {
  botId: string;
  code: string;
}

export interface AdminStats {
  totalUsers: number;
  totalBots: number;
  totalRevenue: number;
  totalMessages: number;
  platformPayments: number;
  botPayments: number;
  completedPayments: number;
}

export interface AdminUser {
  _id: string;
  phone: string;
  platformUnlocked: boolean;
  unlockedBy?: string;
  accessCode?: string | null;
  botAccessCodes?: BotAccessCodeEntry[];
  unlockedBotsCount: number;
  revenue: number;
  messageCount: number;
  joinedAt?: string;
}

export interface AdminUsersResult {
  users: AdminUser[];
  total: number;
  page: number;
  limit: number;
}

export type AdminRevenueResultChartItem = {
  date: string;
  amount: number;
};

export interface AdminRevenueResult {
  chart: AdminRevenueResultChartItem[];
}

export interface Conversation {
  botId: string;
  botName: string;
  botAvatar: string;
  messageCount: number;
  lastMessage: string;
  lastMessageAt: string;
}

export interface ChatMessage {
  _id: string;
  botId: string;
  phone: string;
  role: string;
  text: string;
  createdAt: string;
}

export interface ChatMessageInput {
  phone: string;
  botId: string;
  text: string;
}

export type FetchSessionParams = {
  phone: string;
};

export type FetchPaymentStatusParams = {
  phone: string;
  botId?: string;
  checkoutRequestId?: string;
};

export type FetchPaymentHistoryParams = {
  phone: string;
  page?: number;
  limit?: number;
};

export type FetchConversationsParams = {
  phone: string;
};

export type FetchChatHistoryParams = {
  botId: string;
  phone: string;
};

export type FetchAdminUsersParams = {
  page?: number;
  limit?: number;
};