# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChatBot Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/session:
    get:
      operationId: fetchSession
      tags: [auth]
      summary: Get or create session for a user
      parameters:
        - in: query
          name: phone
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
  /auth/verify:
    post:
      operationId: verifyAccessCode
      tags: [auth]
      summary: Verify platform access code for admin-unlocked users (single-use)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyCodeResult"
        "401":
          description: Invalid code
  /auth/verify-bot:
    post:
      operationId: verifyBotAccessCode
      tags: [auth]
      summary: Verify bot-level access code (single-use, grants bot access)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyBotCodeInput"
      responses:
        "200":
          description: Bot access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyBotCodeResult"
        "401":
          description: Invalid code
...
  /admin/stats:
    get:
      operationId: fetchAdminStats
      tags: [bots]
      summary: Admin — platform statistics
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
  /admin/users:
    get:
      operationId: fetchAdminUsers
      tags: [bots]
      summary: Admin — list users
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResult"
        "401":
          description: Unauthorized
  /admin/revenue:
    get:
      operationId: fetchAdminRevenue
      tags: [bots]
      summary: Admin — revenue chart (last 30 days)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Revenue chart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRevenueResult"
        "401":
          description: Unauthorized
  /admin/users/{phone}/unlock:
    post:
      operationId: adminUnlockUser
      tags: [bots]
      summary: Admin — manually unlock a user and generate access code
      parameters:
        - in: path
          name: phone
          required: true
          schema:
            type: string
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUnlockResult"
        "401":
          description: Unauthorized
  /admin/users/{phone}/bots/{botId}/unlock:
    post:
      operationId: adminUnlockUserBot
      tags: [bots]
      summary: Admin — grant a user free access to a specific bot via access code
      parameters:
        - in: path
          name: phone
          required: true
          schema:
            type: string
        - in: path
          name: botId
          required: true
          schema:
            type: string
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot access code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBotUnlockResult"
        "401":
          description: Unauthorized
        "404":
          description: Bot not found
...
components:
  schemas:
...
    VerifyCodeInput:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
        code:
          type: string
    VerifyCodeResult:
      type: object
      required: [valid, phone, platformUnlocked, unlockedBots]
      properties:
        valid:
          type: boolean
        phone:
          type: string
        platformUnlocked:
          type: boolean
        unlockedBots:
          type: array
          items:
            type: string
        sessionToken:
          type: string
    VerifyBotCodeInput:
      type: object
      required: [phone, botId, code]
      properties:
        phone:
          type: string
        botId:
          type: string
        code:
          type: string
    VerifyBotCodeResult:
      type: object
      required: [valid, phone, botId, platformUnlocked, unlockedBots]
      properties:
        valid:
          type: boolean
        phone:
          type: string
        botId:
          type: string
        platformUnlocked:
          type: boolean
        unlockedBots:
          type: array
          items:
            type: string
        sessionToken:
          type: string
    UserSession:
      type: object
      required: [phone, platformUnlocked, unlockedBots, hasAccessCode]
      properties:
        phone:
          type: string
        platformUnlocked:
          type: boolean
        unlockedBots:
          type: array
          items:
            type: string
        hasAccessCode:
          type: boolean
...
    AdminUnlockResult:
      type: object
      required: [phone, accessCode, message]
      properties:
        phone:
          type: string
        accessCode:
          type: string
        message:
          type: string
    AdminBotUnlockResult:
      type: object
      required: [phone, botId, accessCode, message]
      properties:
        phone:
          type: string
        botId:
          type: string
        accessCode:
          type: string
        message:
          type: string
    BotAccessCodeEntry:
      type: object
      required: [botId, code]
      properties:
        botId:
          type: string
        code:
          type: string
    AdminStats:
      type: object
      required: [totalUsers, totalBots, totalRevenue, totalMessages, platformPayments, botPayments, completedPayments]
      properties:
        totalUsers:
          type: integer
        totalBots:
          type: integer
        totalRevenue:
          type: number
        totalMessages:
          type: integer
        platformPayments:
          type: integer
        botPayments:
          type: integer
        completedPayments:
          type: integer
    AdminUser:
      type: object
      required: [_id, phone, platformUnlocked, unlockedBotsCount, revenue, messageCount]
      properties:
        _id:
          type: string
        phone:
          type: string
        platformUnlocked:
          type: boolean
        unlockedBy:
          type: string
        accessCode:
          type: string
          nullable: true
        botAccessCodes:
          type: array
          items:
            $ref: "#/components/schemas/BotAccessCodeEntry"
        unlockedBotsCount:
          type: integer
        revenue:
          type: number
        messageCount:
          type: integer
        joinedAt:
          type: string
    AdminUsersResult:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminRevenueResult:
      type: object
      required: [chart]
      properties:
        chart:
          type: array
          items:
            type: object
            required: [date, amount]
            properties:
              date:
                type: string
              amount:
                type: number
...