# lib/api-zod/src/generated/types/verifyCodeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatBot Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyCodeInput {
  phone: string;
  code: string;
}